requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
faiss-cpu
torch
